venv_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=py_0
attrs=20.1.0=py_0
black=19.10b0=py_0
blacs=3.0.0=py_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38_0
click=7.1.2=py_0
cycler=0.10.0=py38_0
desktop-app=0.2.6=py38ha925a31_0
distlib=0.2.9=py38_0
freetype=2.10.2=hd328e21_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=1.7.0=py38_0
intel-openmp=2020.2=254
jpeg=9b=hb83a4c4_2
kiwisolver=1.2.0=py38h74a9793_0
labscript=3.0.0=py_0
labscript-c-extensions=1.0.1=py38ha925a31_0
labscript-devices=3.0.0=py_0
labscript-suite=3.0.0=py_0
labscript-utils=3.0.0=py_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
lyse=3.0.0=py_0
lz4-c=1.9.2=h62dcd97_1
mako=1.1.3=pypi_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py38hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msl-loadlib=0.7.0=pypi_0
mypy_extensions=0.4.3=py38_0
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
olefile=0.46=py_0
openssl=1.1.1h=he774522_0
pandas=1.1.1=py38ha925a31_0
pathspec=0.7.0=py_0
pdoc3=0.9.1=pypi_0
pillow=7.2.0=py38hcc1f983_0
pip=20.2.4=py38_0
pydaqmx=1.4.3=py_0
pynivision=2015.0.0=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
pyqtgraph=0.11.0rc0=py_0
pyreadline=2.1=py38_1
pyserial=3.4=py_2
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
pytz=2020.1=py_0
pywin32=227=py38he774522_1
pyzmq=19.0.1=py38ha925a31_1
qt=5.9.7=vc14h73c81de_0
qtutils=3.0.0=py_0
regex=2020.7.14=py38he774522_0
runmanager=3.0.0=py_0
runviewer=3.0.0=py_0
scipy=1.5.0=py38h9439919_0
setuptools=49.6.0=py38_0
setuptools-scm=4.1.2=py_0
setuptools_scm=4.1.2=0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
spinapi=3.2.2=py_0
sqlite=3.33.0=h2a8f88b_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
tornado=6.0.4=py38he774522_1
typed-ast=1.4.1=py38he774522_0
typing_extensions=3.7.4.3=py_0
tzlocal=2.1=py38_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=py_0
wincertstore=0.2=py38_0
windows-curses=2.1.0=py38ha925a31_0
xz=5.2.5=h62dcd97_0
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zprocess=2.20.1=py38_0
zstd=1.4.5=h04227a9_0
